Remove todo about using sentence in Clause id
The Clause id is used in comparisons, so several comparisons in the
Prover have been changed to preserve the semantics of 'compare clauses
on their head and body'. This makes sure that we ignore the sentence in
those places where we must do so.

Alternatively we could have changed the Prover to simplify sentences
based on the context. The choice was made not to do this because this
would lead to an base Prover class with unclear semantics or
responsibilities.